Fix bug where pauseOnReturn was defaulting to true.
The reason is that localStorage only stores strings. So:
> localStorage['bar'] = true;
true
> localStorage['bar']
"true"

Then when localStorage["NoHuluAds:pauseOnReturn"] was set to false it would return "false" which evaluates as a string to true, so pauseOnReturn is set to "false" and not (boolean) false*. Later any check on pauseOnReturn also evaluates to true since it's a string.

Instead, doing var pauseOnReturn = localStorage["NoHuluAds:pauseOnReturn"] == "true"; will be true if it was saved as (boolean) true and (boolean) false if it was set to false or never set at all.

The second line is just to make sure I don't set pauseOnReturn to a string. (though it doesn't seem to be doing it)

* > Boolean("false")
true